fix Admission tests by creating necessary users and approving on their behalf
